Refactor CF parser, update model, fix tests
The model was updated so that ComponentNode uses Sets for its relationships instead of arrays.
The CF parser was split into several classes that deal with nodes in the cloudformation template (resources, parameters, etc).
Nested stacks now have a Component and a ComponentNode (for structure and dependency) to cover the case where a resource depends on the stack directly. Relationship originating in expressions now have their type as the full path of keys to reach the reference.